chatgpt/c02_image_gen/requirements.txt
requests
pyyaml
# fastapi and uvicorn can be uncommented when adding the API backend
# fastapi
# uvicorn
